+++ packages/ui/src/core/drawer/drawer-footer.component.tsx
@@ -1,16 +1,18 @@
-import React from 'react';
+import React, { useContext } from 'react';
 
 import { intrinsicComponent } from '../../utils/functions';
+import DrawerContext from './drawer.context';
 import type { DrawerProps } from './drawer.props';
 import Styled from './drawer.styles';
 
-const DrawerFooter = intrinsicComponent<DrawerProps, HTMLDivElement>(
-  ({ children, ...rest }, ref): JSX.Element => (
-    <Styled.Footer {...rest} ref={ref}>
+const DrawerFooter = intrinsicComponent<DrawerProps, HTMLDivElement>(({ children, ...rest }, ref): JSX.Element => {
+  const { isCollapsed } = useContext(DrawerContext);
+  return (
+    <Styled.Footer isCollapsed={isCollapsed} {...rest} ref={ref}>
       {children}
     </Styled.Footer>
-  )
-);
+  );
+});
 
 DrawerFooter.defaultProps = {};