+++ packages/ui/src/core/drawer/drawer-list.component.tsx
@@ -1,16 +1,18 @@
-import React from 'react';
+import React, { useContext } from 'react';
 
 import { intrinsicComponent } from '../../utils/functions';
+import DrawerContext from './drawer.context';
 import type { DrawerListProps } from './drawer.props';
 import Styled from './drawer.styles';
 
-const DrawerList = intrinsicComponent<DrawerListProps, HTMLUListElement>(
-  ({ children, ...rest }, ref): JSX.Element => (
-    <Styled.List {...rest} ref={ref}>
+const DrawerList = intrinsicComponent<DrawerListProps, HTMLUListElement>(({ children, ...rest }, ref): JSX.Element => {
+  const { isCollapsed } = useContext(DrawerContext);
+  return (
+    <Styled.List isCollapsed={isCollapsed} {...rest} ref={ref}>
       {children}
     </Styled.List>
-  )
-);
+  );
+});
 
 DrawerList.defaultProps = {};