+++ packages/ui/src/core/menu/menu.props.d.ts
@@ -25,5 +25,4 @@ export interface MenuProps extends PropsWithChildren, React.HTMLAttributes<HTMLD
   scroll?: boolean;
   enableUnderlayingEvent?: boolean;
   popperWrapperStyles?: object;
-  applyTransition?: boolean;
 }