+++ packages/ui/src/core/popper/popper.component.tsx
@@ -24,7 +24,7 @@ const Popper = intrinsicComponent<PopperProps, HTMLDivElement>(
       zIndex = 1300,
       enableUnderlayingEvent,
       wrapperStyles = {},
-      applyTransition,
+      ...rest
     }: PopperProps,
     ref
   ): JSX.Element => {
@@ -102,7 +102,7 @@ const Popper = intrinsicComponent<PopperProps, HTMLDivElement>(
     const render = (): JSX.Element => (
       <Styled.PopperWrapper zIndex={zIndex} style={{ ...wrapperStyles }}>
         {overlay && renderOverlay()}
-        <Styled.Popper ref={handlePopperRef} applyTransition={applyTransition}>
+        <Styled.Popper ref={handlePopperRef} {...rest}>
           {children}
           {arrow && (
             <Styled.Arrow
@@ -157,7 +157,6 @@ export const propTypes = {
   zIndex: PT.number,
   enableUnderlayingEvent: PT.bool,
   wrapperStyles: PT.object,
-  applyTransition: PT.bool,
 };
 
 Popper.propTypes = propTypes;