+++ packages/ui/src/core/textarea/textarea.props.d.ts
@@ -13,4 +13,6 @@ export interface TextareaProps extends React.HTMLAttributes<HTMLTextAreaElement>
   readOnly?: boolean;
   disabled?: boolean;
   fullWidth?: boolean;
+  cols?: number;
+  rows?: number;
 }