+++ packages/ui/stories/core/drawer.story.tsx
@@ -122,7 +122,7 @@ const BasicTemplate: Story<DrawerProps> = ({ ...args }) => {
   };
 
   return (
-    <>
+    <div>
       <StyledHeader>
         <IconButton
           style={{
@@ -138,7 +138,16 @@ const BasicTemplate: Story<DrawerProps> = ({ ...args }) => {
         </IconButton>
         Drawer
       </StyledHeader>
-      <Drawer {...args} open={isDrawerOpened} onClose={() => setIsDrawerOpened(false)} onCollapse={onCollapse}>
+      <Drawer
+        {...args}
+        open={isDrawerOpened}
+        collapsed={isCollapsed}
+        onClose={() => setIsDrawerOpened(false)}
+        onCollapse={onCollapse}
+        style={{
+          marginTop: '46px',
+        }}
+      >
         <DrawerHeader>
           <IconButton
             style={{
@@ -190,7 +199,7 @@ const BasicTemplate: Story<DrawerProps> = ({ ...args }) => {
           </DrawerItemButton>
         </DrawerFooter>
       </Drawer>
-    </>
+    </div>
   );
 };